% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_effects.R
\name{ComputeEffectsLinear}
\alias{ComputeEffectsLinear}
\title{Computes direct and indirect effects for linear models.}
\usage{
ComputeEffectsLinear(fit)
}
\arguments{
\item{fit}{List with path coefficients.}
}
\value{
Matrix of direct, indirect, and total effects for each xx.
}
\description{
\code{ComputeEffectsLinear} computes the direct and indirect effects
for linear models. These are simple functions of the path coefficients.
Argument can be path coefficients fitted to data, in format output by
\code{ComputePath} or simulation parameters output by function such as
\code{SimpleSim}.
}
\examples{
set.seed(1234)
params <- SimpleSim()
dat <- SimulateData(params)
fit <- ComputePath(dat)
ComputeEffectsLinear(params) ## exact
ComputeEffectsLinear(fit) ## based on sample
}
