% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_effects.R
\name{ComputeEffectxx}
\alias{ComputeEffectxx}
\title{Computes direct, indirect, or total effects of xx}
\usage{
ComputeEffectxx(dat, fit, effect, xp = 0, xpp = 1, risk_scale = NULL,
  mmn = FALSE, rmean = NULL)
}
\arguments{
\item{dat}{List containing data (see SimulateData for format)}

\item{fit}{List with path coefficients.}

\item{effect}{Either 'direct', 'indirect', or 'total'}

\item{xp}{The initial value of x.}

\item{xpp}{The new value of x.}

\item{risk_scale}{Either 'diff' (for risk difference) or 'ratio' (for odds).
see details for more information.}

\item{mmn}{Should gene set network be simulated. If FALSE assume
conditionally independent mediators.}
}
\value{
Vector of effects.
}
\description{
\code{ComputeEffectxx} computes either direct, indirect, or total effect
of each xx variable on response y when transition x from xp to xpp.
}
\examples{
params <- SimpleSim()
dat <- SimulateData(params)
fit <- ComputePath(dat)
d <- ComputeEffectxx(dat,fit,"direct")
i <- ComputeEffectxx(dat,fit,"indirect")
to <- ComputeEffectxx(dat,fit,"total")
## effects computed by estimating integrals
print(c(d,i,to))
## effects using model linearity assumption
ComputeEffectsLinear(fit) ## based on sample
## exact effects using simulation coefficients
ComputeEffectsLinear(params) ## exact
}
