% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_effects.R
\name{ComputePath}
\alias{ComputePath}
\title{Computes path coefficients.}
\usage{
ComputePath(dat, reg = FALSE, mmn = FALSE)
}
\arguments{
\item{dat}{List containing data (see SimulateData for format)}

\item{reg}{Should path coefficient estimates be regularized. Experimental.
Defaults to FALSE.}

\item{mmn}{Should residuals from mm | xx regression be returned.
If FALSE return NULL. Must be true if estimating direct/indirect effects
with graph structure.}
}
\value{
List containing path coefficients, variance estimates, residuals.
}
\description{
\code{ComputePath} computes the coefficients for for all paths linking
response, mediators, treatment and potential confounders.
}
\examples{
params <- SimpleSim()
dat <- SimulateData(params)
fit <- ComputePath(dat)
print(fit$xx_direct)
print(params$xx_direct)
}
