% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{QuickSimMultipleMediator}
\alias{QuickSimMultipleMediator}
\title{Generate simulation parameters assuming a simple path structure}
\usage{
QuickSimMultipleMediator(n, nmm, family, var_mm = 10, xx_direct = 0.5,
  mm_direct = 0.5, xx_prob = 0.2)
}
\arguments{
\item{n}{Sample Size}

\item{nmm}{Number of mediators}

\item{family}{Response type}

\item{var_mm}{Variance of the mediators}

\item{xx_direct}{Path coefficient form x to response}

\item{mm_direct}{Path coefficient from mediators to response}

\item{xx_prob}{Probability x is 1.}
}
\value{
List that the parameterizes simulation. Typically used as
  argument in subsequent call to \code{SimulateData}
}
\description{
\code{SimpleSim} creates simulation parameters which are then used
  to generate data with \code{SimulateData}. Mostly used for testing.
}
\examples{
params <- QuickSimMultipleMediator(500,5,"cox")
names(params)
}
