% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{SimpleSim}
\alias{SimpleSim}
\title{Generate simulation parameters.}
\usage{
SimpleSim(n = 500, family = "gaussian")
}
\arguments{
\item{n}{Sample Size}

\item{family}{Either 'gaussian', 'binomial', or 'cox' depending on
desired response model.}
}
\value{
List that the parameterizes simulation. Typically used as
  argument in subsequent call to \code{SimulateData}
}
\description{
\code{SimpleSim} creates simulation parameters which are then used
  to generate data with \code{SimulateData}. Mostly used for testing.
}
\examples{
params <- SimpleSim()
names(params)
}
